+++ sbin/ifconfig/ifmac.c
@@ -31,7 +31,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $FreeBSD: release/7.0.0/sbin/ifconfig/ifmac.c 139494 2004-12-31 19:46:27Z sam $
+ * $FreeBSD$
  */
 
 #include <sys/param.h>
@@ -112,7 +112,7 @@ static __constructor void
 mac_ctor(void)
 {
 #define	N(a)	(sizeof(a) / sizeof(a[0]))
-	int i;
+	size_t i;
 
 	for (i = 0; i < N(mac_cmds);  i++)
 		cmd_register(&mac_cmds[i]);
